requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pycryptodome==3.22.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
black==25.1.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mypy-extensions==1.0.0
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pycryptodome==3.22.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
